@@ -16,27 +16,27 @@
   "author": "Conveyal",
   "license": "MIT",
   "dependencies": {
-    "color": "^0.11.3",
-    "debug": "^2.2.0",
-    "jsolines": "^0.2.2",
-    "lodash.fill": "^3.3.2",
-    "lodash.slice": "^4.0.2",
+    "color": "^0.11.4",
+    "debug": "^2.3.3",
+    "jsolines": "^0.2.3",
+    "lodash.fill": "^3.4.0",
+    "lodash.slice": "^4.2.0",
     "lonlng": "^0.2.0",
     "web-worker-promise-interface": "^0.2.0"
   },
   "devDependencies": {
     "concat-stream": "^1.5.2",
-    "isomorphic-fetch": "^2.2.0",
+    "isomorphic-fetch": "^2.2.1",
     "leaflet-transitivelayer": "^0.2.0",
     "lodash.reverse": "^4.0.1",
-    "mapbox.js": "^2.2.4",
+    "mapbox.js": "^3.0.1",
     "mastarm": "^3.0.0",
     "normalize.css": "^5.0.0",
-    "react": "^15.3.1",
-    "react-dom": "^15.3.1",
-    "semantic-release": "^4.3.5",
-    "tape": "^4.2.2",
-    "transitive-js": "^0.9.0"
+    "react": "^15.4.1",
+    "react-dom": "^15.4.1",
+    "semantic-release": "^6.3.2",
+    "tape": "^4.6.3",
+    "transitive-js": "^0.9.2"
   },
   "standard": {
     "parser": "babel-eslint"
