@@ -34,13 +34,24 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
    * before calling this function if you expect to use it elsewhere
    */
   putGrid (key, grid) {
-    this.work({
+    return this.work({
       command: 'putGrid',
       message: { key, grid }
     })
   }
 
-  drawTile (canvas, {x, y}, zoom) {
+  createTile = (coords, done) => {
+    const tile = document.createElement('canvas')
+    tile.height = imageHeight
+    tile.width = imageWidth
+    console.log(coords)
+    this.drawTile(tile, coords, coords.z)
+      .then(() => done(null, tile))
+      .catch((err) => done(err, tile))
+    return tile
+  }
+
+  drawTile = (canvas, {x, y}, zoom) => {
     const scaleFactor = Math.pow(2, zoom - this.query.zoom)
 
     // find top-left coords at zoom 10
@@ -52,7 +63,7 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
     const ctx = canvas.getContext('2d')
     const imageData = ctx.createImageData(imageHeight, imageWidth)
 
-    this.work({
+    return this.work({
       command: 'drawTile',
       message: {
         imageData,
@@ -60,8 +71,9 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
         xoffset,
         yoffset
       }
-    }).then(message => {
+    }).then((message) => {
       ctx.putImageData(message, 0, 0)
+      return message
     })
   }
 
